% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root-escape.R
\name{rootexpr}
\alias{rootexpr}
\title{ROOT escaping.}
\usage{
rootexpr(x)
}
\description{
These functions are critical when writing functions that translate R
functions to ROOT functions. Typically a conversion function should escape
all it's inputs and return an sql object.
}
