% Generated by roxygen2 (4.0.2): do not edit by hand
\name{src_root}
\alias{src_root}
\alias{tbl.src_root}
\alias{tbl_rootchain}
\title{Connect to a ROOT file.}
\usage{
src_root(path, ...)

\method{tbl}{src_root}(src, tree_name, ...)

tbl_rootchain(tbls, tree_name)
}
\arguments{
\item{path}{Path to ROOT file.}

\item{...}{Included for compatibility with the generic, but otherwise
ignored.}

\item{src}{A ROOT src created with \code{src_root}.}

\item{tree_name}{Name (and path) to the tree in the ROOT file.
Vector is recycled if used in \code{tbl_rootchain}.}

\item{tbls}{Either a list of ROOT trees created with \code{tbl}, or
  a character vector of file names.}
}
\description{
Use \code{src_root} to open an existing ROOT file,
and \code{tbl} to open trees within that ROOT file.
Use \code{tbl_rootchain} to open a "chain", i.e. bind
multiple trees from one or many files.
}

