% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-root.R
\name{translate_root}
\alias{translate_root}
\alias{translate_root_q}
\title{Translate an expression to ROOT's TTree::Draw syntax.}
\usage{
translate_root(
  ...,
  tbl = NULL,
  env = parent.frame(),
  variant = NULL,
  window = FALSE
)

translate_root_q(expr, tbl = NULL, env = parent.frame())
}
\description{
Translate an expression to ROOT's TTree::Draw syntax.
}
